Decided to merge the books by the page. Hopefully, it should work all the same
